% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-moves.R
\name{altitude_change_opts}
\alias{altitude_change_opts}
\title{if soaring can land, if grounded can take off}
\usage{
altitude_change_opts(.l)
}
\description{
if soaring can land, if grounded can take off
}
\keyword{internal}
