% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-game.R
\name{bpl}
\alias{bpl}
\title{Build Player List}
\usage{
bpl(.name, .starting_loc, .units)
}
\description{
Build the list that can be passed in (as part of a list)
to \code{\link[=new_game]{new_game()}} \code{players} argument.
}
