% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{build_name}
\alias{build_name}
\alias{wu_epithet}
\title{Build a name for a new unit}
\usage{
build_name(force_org = "any")

wu_epithet(.n)
}
\arguments{
\item{force_org}{Specify a \code{force_org} slot for tailored name. If not passed,
will choose at random.}
}
\description{
Build a name for a new unit
}
\section{Functions}{
\itemize{
\item \code{wu_epithet}: Pass a name to this to get a Wu adjective appended.
}}

