% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir_empty}
\alias{create_dir_empty}
\title{Create an empty copy of a directory, deleting one that's there if necessary}
\usage{
create_dir_empty(.path)
}
\description{
Create an empty copy of a directory, deleting one that's there if necessary
}
\keyword{internal}
