% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_create_empty}
\alias{dir_create_empty}
\title{Create an empty copy of a directory, deleting one that's there if necessary}
\usage{
dir_create_empty(.path)
}
\description{
Create an empty copy of a directory, deleting one that's there if necessary
}
\keyword{internal}
