% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-map.R
\name{draw_map}
\alias{draw_map}
\title{Draw the game map}
\usage{
draw_map(game, .p = NULL)
}
\arguments{
\item{game}{The game object that contains the map etc}
}
\value{
Returns a ggplot object of the map
}
\description{
Draw the game map
}
