% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake-db.R
\name{game_db_path}
\alias{game_db_path}
\alias{game_json_path}
\alias{game_csv_path}
\alias{game_map_path}
\alias{game_img_path}
\title{Get path to game db files on disk}
\usage{
game_db_path(game, ext = NULL, create = FALSE)

game_json_path(game_name)

game_csv_path(game)

game_map_path(game)

game_img_path(game, .p)
}
\arguments{
\item{game}{game object}

\item{ext}{extension, if NULL returns dir}

\item{create}{if TRUE, create it. Only works for dir.}
}
\description{
Get path to game db files on disk
}
\section{Functions}{
\itemize{
\item \code{game_json_path}: Path to general game info json

\item \code{game_csv_path}: Path to map_df csv

\item \code{game_map_path}: Path to map json

\item \code{game_img_path}: Path to map json
}}

\keyword{internal}
