% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake-db.R
\name{game_json_to_disk}
\alias{game_json_to_disk}
\title{Write general game info json to disk}
\usage{
game_json_to_disk(game)
}
\description{
Idea is this is immutable stuff created at new_game().
Is that right, or do we write this every turn (if it can change)?
}
