% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake-db.R
\name{game_path}
\alias{game_path}
\title{Takes a game name and returns the path to a relevant file}
\usage{
game_path(game_name, ext = "")
}
\description{
Takes a game name and returns the path to a relevant file
}
\keyword{internal}
