% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-helpers.R
\name{game_script_from_rmd}
\alias{game_script_from_rmd}
\title{Render a game Rmd into a game R script}
\usage{
game_script_from_rmd(game_name, run = FALSE)
}
\arguments{
\item{game_name}{name of the game. Will look for a dir in
\code{file.path(getOption(MC.games_dir), sanitize_name(game_name))} containing
game files}
}
\description{
Render a game Rmd into a game R script
}
