% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unit_attrs}
\alias{get_unit_attrs}
\title{See tibble of unit attributes}
\usage{
get_unit_attrs(.u)
}
\arguments{
\item{.u}{Character vector of unit types}
}
\description{
See tibble of unit attributes
}
\keyword{internal}
