% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake-db.R
\name{increment_turn_phase}
\alias{increment_turn_phase}
\title{Increment the game turn/phase}
\usage{
increment_turn_phase(game)
}
\description{
If no conflicts, increment to new turn.
If conflicts, increment phase.
}
\keyword{internal}
