% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-moves.R
\name{input_moves}
\alias{input_moves}
\alias{input_player_code}
\alias{input_unit}
\alias{input_action}
\alias{input_loc}
\title{Input moves for one (group of) units.}
\usage{
input_player_code(game, player_code)

input_unit(game, .p, .u)

input_action(unit_df)

input_loc(unit_df, game)
}
\arguments{
\item{unit_df}{The tibble of units that's acting}
}
\description{
Input moves for one (group of) units.
}
\section{Functions}{
\itemize{
\item \code{input_player_code}: Input player code to trigger move. Returns the name of the player.

\item \code{input_unit}: Takes player and .u and returns legal actions.

\item \code{input_action}: Takes units (from \code{input_units()}) and adds the chosen action.
Returns the \code{unit_df} with the chosen action filled in.

\item \code{input_loc}: Takes units with action (from \code{input_action()}) and adds the loc(s)
Returns the \code{unit_df} with the chosen locs filled in.
}}

