% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-unit.R
\name{loc_to_unit}
\alias{loc_to_unit}
\title{check for loc passed as .u and convert to units in loc}
\usage{
loc_to_unit(game, .p, .u)
}
\description{
check for loc passed as .u and convert to units in loc
}
\keyword{internal}
