% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-new-turn.R
\name{make_new_move}
\alias{make_new_move}
\title{Actually make the move}
\usage{
make_new_move(game, .p, .u, .a = NULL, .l = NULL)
}
\arguments{
\item{game}{\code{MC_game} object to modify}

\item{.p}{String; player name}

\item{.u}{units; passed to input_units()}

\item{.a}{action to take. If \code{NULL}, the default, ask the user.}

\item{.l}{location to act. If \code{NULL}, the default, ask the user.}
}
\value{
The modified \code{MC_game} object passed to \code{game}
}
\description{
Actually make the move
}
