% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-unit.R
\name{modify_unit}
\alias{modify_unit}
\title{Modify the action of a unit or group of units}
\usage{
modify_unit(game, .p, .u, .a, .l)
}
\arguments{
\item{game}{The game to modify the unit(s) in}

\item{.p}{The player who owns the unit(s)}

\item{.u}{Either a character vector of unit names to modify, a vector of
unit_id's , or a scalar of a map loc, in which case all units in that loc
will be modified.}

\item{.a}{Action to assign to units}

\item{.l}{Location to move units to (or same as where they are if not moving)}
}
\description{
Modify the action of a unit or group of units
}
