% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MundusCentrum.R
\name{MundusCentrum}
\alias{MundusCentrum}
\title{Mundus Centrum

An open source long form strategy, tactics, and logistics war game, heavily
inspired by Warhammer 40k and the blog of Brett Devereux.}
\description{
Mundus Centrum

An open source long form strategy, tactics, and logistics war game, heavily
inspired by Warhammer 40k and the blog of Brett Devereux.
}
