% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-game.R
\name{new_game}
\alias{new_game}
\title{Start a new Mundus Centrum game}
\usage{
new_game(name, players, points = NULL)
}
\arguments{
\item{name}{Name of game}

\item{players}{List of players. See "Details" for structure of list.}

\item{points}{Total number of points each player starts with. If left \code{NULL},
will look for a \code{points} element in each player data element.}
}
\description{
Start a new Mundus Centrum game
}
\details{
\strong{Player data format} -- Each element of the list passed to \code{players} must
have the following elements:
\itemize{
\item \code{name} -- Player name. Must be unique amongst players.
\item \code{team} -- One of the legal teams. See list with \code{teams()}.
\item \code{units} -- Path to a units.csv file to load.
\item \code{points} (optional) -- how many points to start the game with. Overriden if
anything is passed to \code{points} argument of the function.
}
}
