% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile-map.R
\name{players_to_global_map}
\alias{players_to_global_map}
\title{Reconcile player maps with global map}
\usage{
players_to_global_map(game)
}
\description{
Reconcile player maps with global map
}
\keyword{internal}
