% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-helpers.R
\name{print_map_df}
\alias{print_map_df}
\title{Print list of units on the map}
\usage{
print_map_df(game, .p = NULL)
}
\description{
Calls \code{\link[=get_player_map]{get_player_map()}} and formats for printing
}
