% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-helpers.R
\name{print_shiny_df}
\alias{print_shiny_df}
\title{Return the map_df for displaying in shiny sidebar}
\usage{
print_shiny_df(game, .p = NULL)
}
\description{
Return the map_df for displaying in shiny sidebar
}
