% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-game.R
\name{publish_game_html}
\alias{publish_game_html}
\title{Copy game HTML files to a new dir (for publishing and hosting)}
\usage{
publish_game_html(game_name, dest_dir, overwrite = FALSE)
}
\description{
Copy game HTML files to a new dir (for publishing and hosting)
}
