% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile-player-orders.R
\name{reconcile_player_orders}
\alias{reconcile_player_orders}
\title{Reconcile player maps with global map}
\usage{
reconcile_player_orders(game)
}
\description{
Run after players have updated their maps. This will add all unit placements
on the player maps to the global map. If there are opposing units in any
locations, it will return a tibble of the units in conflict zones and warn
about that the conflict must be resolved. If there are no conflicts,
invisibly returns the tibble with all units on the board.
}
