% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-game.R
\name{render_game}
\alias{render_game}
\title{Render the Rmd files for a game}
\usage{
render_game(game_path, players = FALSE, html = FALSE)
}
\description{
Takes an \code{.Rmd} template and a \code{.json} as input and renders
the markdown(s)
}
