% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-unit.R
\name{unit_id_to_name}
\alias{unit_id_to_name}
\title{check for unit_id passed as .u and convert to unit_name}
\usage{
unit_id_to_name(game, .p, .ux)
}
\description{
check for unit_id passed as .u and convert to unit_name
}
\keyword{internal}
