% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-army-from-csv.R
\name{add_army_from_csv}
\alias{add_army_from_csv}
\title{Create csv needed to start game from an input csv}
\usage{
add_army_from_csv(in_path, out_path, starting_loc, player_name = NULL)
}
\arguments{
\item{in_path}{Path to input CSV}

\item{out_path}{Path to write army CSV to}

\item{starting_loc}{Map location to start units in}

\item{player_name}{Optional string, if passed with filter input df with \code{filter(.data$player == player_name)}}
}
\description{
Create csv needed to start game from an input csv
}
