% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-name.R
\name{build_name}
\alias{build_name}
\alias{show_me_names}
\alias{wu_epithet}
\title{Build a name for a new unit}
\usage{
build_name(.u = NULL)

show_me_names(.n = 8)

wu_epithet(.n)
}
\arguments{
\item{.u}{Specify a \code{unit_type} for tailored name. If not passed,
will choose at random.}

\item{.n}{How many names you wanna see?}
}
\description{
Build a name for a new unit
}
\section{Functions}{
\itemize{
\item \code{show_me_names}: Show some sample names you could build.

\item \code{wu_epithet}: Pass a name to this to get a Wu adjective appended.
}}

