% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-map.R
\name{draw_map}
\alias{draw_map}
\title{Draw the game map}
\usage{
draw_map(game_df)
}
\arguments{
\item{game_df}{Takes a tibble output from \code{\link[=reconcile_player_orders]{reconcile_player_orders()}}}
}
\value{
Returns a ggplot object of the map
}
\description{
Draw the game map
}
