% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile-player-orders.R
\name{players_to_global_map}
\alias{players_to_global_map}
\title{Update global map from player maps}
\usage{
players_to_global_map(game)
}
\description{
Modifies the global map by
}
\keyword{internal}
